refactor: smsclient, dto 패키지 domain 패키지로 이동
