+++ src/main/java/com/gloddy/server/authSms/domain/SmsClient.java
@@ -1,6 +1,6 @@
-package com.gloddy.server.authSms;
+package com.gloddy.server.authSms.domain;
 
-import com.gloddy.server.authSms.dto.SmsRequest;
+import com.gloddy.server.authSms.domain.dto.SmsRequest;
 
 public interface SmsClient {
     void send(SmsRequest.Send dto);
