+++ src/main/java/com/gloddy/server/authSms/domain/dto/SmsResponse.java
@@ -4,21 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-
-public class SmsRequest {
-
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class Send {
-        private String receivingNumber;
-    }
-
+public class SmsResponse {
     @Getter
     @NoArgsConstructor
     @AllArgsConstructor
     public static class VerifyCode {
-        private String number;
-        private String code;
+        private boolean isVerify;
     }
 }
